most functionality

+++ vagrant/catalog/project.py
@@ -301,18 +301,19 @@ def allSongsJSON():
 @app.route('/playlist/')
 def showPlaylists():
     playlists = session.query(Playlist).order_by(asc(Playlist.title))
-    songs = session.query(Song).order_by(desc(Song.id)).limit(2)
+    songs = session.query(Song,Playlist,User).filter(Song.playlist_id == Playlist.id, Song.user_id == User.id).order_by(Song.id.desc()).limit(2)
     if 'username' not in login_session:
-        return render_template('public_playlists.html', playlists=playlists)
+        return render_template('public_playlists.html', playlists=playlists, items = songs)
+        #for song in songs:
     else:
         return render_template('playlist.html', playlists=playlists, items = songs)
 
 @app.route('/playlist/new/', methods=['GET', 'POST'])
 def newPlaylist():
-    #if 'username' not in login_session:
-     #   return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
-        newPlaylist = Playlist(title=request.form['title'])#, user_id=login_session['user_id'])
+        newPlaylist = Playlist(title=request.form['title'], description=request.form['description'], user_id=login_session['user_id'])
         session.add(newPlaylist)
         flash('New Playlist %s Successfully Created' % newPlaylist.title)
         session.commit()
@@ -330,16 +331,16 @@ def showSongs(playlist_id):
     if 'username' not in login_session or creator.id != login_session['user_id']:
         return render_template('public_list_of_songs.html', items=items, playlist=playlist, creator=creator)
     else:
-        return render_template('list_of_songs.html', items=items, playlist=playlist)
+        return render_template('list_of_songs.html', items=items, playlist=playlist, creator=creator)
 
 @app.route('/playlist/<int:playlist_id>/edit/', methods=['GET', 'POST'])
 def editPlaylist(playlist_id):
     editedPlaylist = session.query(
         Playlist).filter_by(id=playlist_id).one()
-    #if 'username' not in login_session:
-     #   return redirect('/login')
-    #if editedRestaurant.user_id != login_session['user_id']:
-    #    return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()''>"
+    if 'username' not in login_session:
+        return redirect('/login')
+    if editedPlaylist.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         if request.form['title']:
             editedPlaylist.title = request.form['title']
@@ -353,11 +354,11 @@ def editPlaylist(playlist_id):
 # Create a new menu item
 @app.route('/playlist/<int:playlist_id>/song/new/', methods=['GET', 'POST'])
 def newSong(playlist_id):
-    #if 'username' not in login_session:
-     #   return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     playlist = session.query(Playlist).filter_by(id=playlist_id).one()
-    #if login_session['user_id'] != restaurant.user_id:
-        #return "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.');}</script><body onload='myFunction()''>"
+    if login_session['user_id'] != playlist.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         newSong = Song(title=request.form['title'], performed_by=request.form['performed_by'], album=request.form[
                                'album'], notes=request.form['notes'], playlist_id=playlist_id, user_id=playlist.user_id)
@@ -374,10 +375,10 @@ def deletePlaylist(playlist_id):
     playlistToDelete = session.query(
         Playlist).filter_by(id=playlist_id).one()
     songsToDelete = session.query(Song).filter_by(playlist_id = playlist_id).all()
-    #if 'username' not in login_session:
-     #   return redirect('/login')
-    #if restaurantToDelete.user_id != login_session['user_id']:
-     #   return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()''>"
+    if 'username' not in login_session:
+        return redirect('/login')
+    if playlistToDelete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete this playlist. Please create your own playlist in order to delete.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(playlistToDelete)
         flash('%s Successfully Deleted' % playlistToDelete.title)
@@ -389,12 +390,12 @@ def deletePlaylist(playlist_id):
 # Delete a song
 @app.route('/playlist/<int:playlist_id>/song/<int:song_id>/edit', methods=['GET', 'POST'])
 def editSong(playlist_id, song_id):
-    #if 'username' not in login_session:
-     #   return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     editedSong = session.query(Song).filter_by(id=song_id).one()
     playlist = session.query(Playlist).filter_by(id=playlist_id).one()
-    #if login_session['user_id'] != restaurant.user_id:
-     #   return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.');}</script><body onload='myFunction()''>"
+    if login_session['user_id'] != playlist.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to edit songs in this playlist. Please create your own playlists.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         if request.form['title']:
             editedSong.title = request.form['title']
@@ -414,12 +415,12 @@ def editSong(playlist_id, song_id):
 # Delete a menu item
 @app.route('/playlist/<int:playlist_id>/song/<int:song_id>/delete', methods=['GET', 'POST'])
 def deleteSong(playlist_id, song_id):
-    #if 'username' not in login_session:
-     #   return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     playlist = session.query(Playlist).filter_by(id=playlist_id).one()
     songToDelete = session.query(Song).filter_by(id=song_id).one()
-    #if login_session['user_id'] != restaurant.user_id:
-     #   return "<script>function myFunction() {alert('You are not authorized to delete menu items to this restaurant. Please create your own restaurant in order to delete items.');}</script><body onload='myFunction()''>"
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to delete songs in this playlist.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(songToDelete)
         session.commit()
@@ -429,201 +430,6 @@ def deleteSong(playlist_id, song_id):
         return render_template('delete_song.html', item=songToDelete, playlist_id = playlist_id)
  
 
-'''
-
-# DISCONNECT - Revoke a current user's token and reset their login_session
-
-
-
-# JSON APIs to view Restaurant Information
-@app.route('/restaurant/<int:restaurant_id>/menu/JSON')
-def restaurantMenuJSON(restaurant_id):
-    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-    items = session.query(MenuItem).filter_by(
-        restaurant_id=restaurant_id).all()
-    return jsonify(MenuItems=[i.serialize for i in items])
-
-
-@app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/JSON')
-def menuItemJSON(restaurant_id, menu_id):
-    Menu_Item = session.query(MenuItem).filter_by(id=menu_id).one()
-    return jsonify(Menu_Item=Menu_Item.serialize)
-
-
-@app.route('/restaurant/JSON')
-def restaurantsJSON():
-    restaurants = session.query(Restaurant).all()
-    return jsonify(restaurants=[r.serialize for r in restaurants])
-
-
-# Show all restaurants
-@app.route('/')
-@app.route('/restaurant/')
-def showRestaurants():
-    restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-    if 'username' not in login_session:
-        return render_template('publicrestaurants.html', restaurants=restaurants)
-    else:
-        return render_template('restaurants.html', restaurants=restaurants)
-
-# Create a new restaurant
-
-
-@app.route('/restaurant/new/', methods=['GET', 'POST'])
-def newRestaurant():
-    if 'username' not in login_session:
-        return redirect('/login')
-    if request.method == 'POST':
-        newRestaurant = Restaurant(
-            name=request.form['name'], user_id=login_session['user_id'])
-        session.add(newRestaurant)
-        flash('New Restaurant %s Successfully Created' % newRestaurant.name)
-        session.commit()
-        return redirect(url_for('showRestaurants'))
-    else:
-        return render_template('newRestaurant.html')
-
-# Edit a restaurant
-
-
-@app.route('/restaurant/<int:restaurant_id>/edit/', methods=['GET', 'POST'])
-def editRestaurant(restaurant_id):
-    editedRestaurant = session.query(
-        Restaurant).filter_by(id=restaurant_id).one()
-    if 'username' not in login_session:
-        return redirect('/login')
-    if editedRestaurant.user_id != login_session['user_id']:
-        return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()''>"
-    if request.method == 'POST':
-        if request.form['name']:
-            editedRestaurant.name = request.form['name']
-            flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
-            return redirect(url_for('showRestaurants'))
-    else:
-        return render_template('editRestaurant.html', restaurant=editedRestaurant)
-
-
-# Delete a restaurant
-@app.route('/restaurant/<int:restaurant_id>/delete/', methods=['GET', 'POST'])
-def deleteRestaurant(restaurant_id):
-    restaurantToDelete = session.query(
-        Restaurant).filter_by(id=restaurant_id).one()
-    if 'username' not in login_session:
-        return redirect('/login')
-    if restaurantToDelete.user_id != login_session['user_id']:
-        return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()''>"
-    if request.method == 'POST':
-        session.delete(restaurantToDelete)
-        flash('%s Successfully Deleted' % restaurantToDelete.name)
-        session.commit()
-        return redirect(url_for('showRestaurants', restaurant_id=restaurant_id))
-    else:
-        return render_template('deleteRestaurant.html', restaurant=restaurantToDelete)
-
-# Show a restaurant menu
-
-
-@app.route('/restaurant/<int:restaurant_id>/')
-@app.route('/restaurant/<int:restaurant_id>/menu/')
-def showMenu(restaurant_id):
-    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-    creator = getUserInfo(restaurant.user_id)
-    items = session.query(MenuItem).filter_by(
-        restaurant_id=restaurant_id).all()
-    if 'username' not in login_session or creator.id != login_session['user_id']:
-        return render_template('publicmenu.html', items=items, restaurant=restaurant, creator=creator)
-    else:
-        return render_template('menu.html', items=items, restaurant=restaurant, creator=creator)
-
-
-# Create a new menu item
-@app.route('/restaurant/<int:restaurant_id>/menu/new/', methods=['GET', 'POST'])
-def newMenuItem(restaurant_id):
-    if 'username' not in login_session:
-        return redirect('/login')
-    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-    if login_session['user_id'] != restaurant.user_id:
-        return "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.');}</script><body onload='myFunction()''>"
-        if request.method == 'POST':
-            newItem = MenuItem(name=request.form['name'], description=request.form['description'], price=request.form[
-                               'price'], course=request.form['course'], restaurant_id=restaurant_id, user_id=restaurant.user_id)
-            session.add(newItem)
-            session.commit()
-            flash('New Menu %s Item Successfully Created' % (newItem.name))
-            return redirect(url_for('showMenu', restaurant_id=restaurant_id))
-    else:
-        return render_template('newmenuitem.html', restaurant_id=restaurant_id)
-
-# Edit a menu item
-
-
-@app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET', 'POST'])
-def editMenuItem(restaurant_id, menu_id):
-    if 'username' not in login_session:
-        return redirect('/login')
-    editedItem = session.query(MenuItem).filter_by(id=menu_id).one()
-    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-    if login_session['user_id'] != restaurant.user_id:
-        return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.');}</script><body onload='myFunction()''>"
-    if request.method == 'POST':
-        if request.form['name']:
-            editedItem.name = request.form['name']
-        if request.form['description']:
-            editedItem.description = request.form['description']
-        if request.form['price']:
-            editedItem.price = request.form['price']
-        if request.form['course']:
-            editedItem.course = request.form['course']
-        session.add(editedItem)
-        session.commit()
-        flash('Menu Item Successfully Edited')
-        return redirect(url_for('showMenu', restaurant_id=restaurant_id))
-    else:
-        return render_template('editmenuitem.html', restaurant_id=restaurant_id, menu_id=menu_id, item=editedItem)
-
-
-# Delete a menu item
-@app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods=['GET', 'POST'])
-def deleteMenuItem(restaurant_id, menu_id):
-    if 'username' not in login_session:
-        return redirect('/login')
-    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-    itemToDelete = session.query(MenuItem).filter_by(id=menu_id).one()
-    if login_session['user_id'] != restaurant.user_id:
-        return "<script>function myFunction() {alert('You are not authorized to delete menu items to this restaurant. Please create your own restaurant in order to delete items.');}</script><body onload='myFunction()''>"
-    if request.method == 'POST':
-        session.delete(itemToDelete)
-        session.commit()
-        flash('Menu Item Successfully Deleted')
-        return redirect(url_for('showMenu', restaurant_id=restaurant_id))
-    else:
-        return render_template('deleteMenuItem.html', item=itemToDelete)
-
-
-# Disconnect based on provider
-@app.route('/disconnect')
-def disconnect():
-    if 'provider' in login_session:
-        if login_session['provider'] == 'google':
-            gdisconnect()
-            del login_session['gplus_id']
-            credentials = login_session.get('credentials')
-            del credentials
-        if login_session['provider'] == 'facebook':
-            fbdisconnect()
-            del login_session['facebook_id']
-        del login_session['username']
-        del login_session['email']
-        del login_session['picture']
-        del login_session['user_id']
-        del login_session['provider']
-        flash("You have successfully been logged out.")
-        return redirect(url_for('showRestaurants'))
-    else:
-        flash("You were not logged in")
-        return redirect(url_for('showRestaurants'))
-
-'''
 if __name__ == '__main__':
     app.secret_key = 'super_secret_key'
     app.debug = True
